Moves partiql-grammar to org.partiql.lang.syntax.antlr so we don't vend a separate jar
