+++ lang/antlr/PartiQL.g4
@@ -1,4 +1,3 @@
-
 grammar PartiQL;
 
 options {
