+++ lang/src/org/partiql/lang/syntax/PartiQLParser.kt
@@ -33,10 +33,9 @@ import org.partiql.lang.errors.PropertyValueMap
 import org.partiql.lang.types.CustomType
 import org.partiql.lang.util.getIonValue
 import org.partiql.lang.util.getPartiQLTokenType
-import org.partiql.lang.visitors.PartiQLVisitor
 import java.nio.charset.StandardCharsets
-import org.partiql.grammar.parser.generated.PartiQLParser as GeneratedParser
-import org.partiql.grammar.parser.generated.PartiQLTokens as GeneratedLexer
+import org.partiql.lang.syntax.antlr.PartiQLParser as GeneratedParser
+import org.partiql.lang.syntax.antlr.PartiQLTokens as GeneratedLexer
 
 /**
  * Extends [Parser] to provide a mechanism to parse an input query string. It internally uses ANTLR's generated parser,
