--- partiql-grammar/README.md
@@ -1,17 +1,0 @@
-# PartiQL Grammar
-
-## Description
-
-This Gradle sub-project is dedicated to providing a generated parser by leveraging [ANTLR](https://docs.gradle.org/current/userguide/antlr_plugin.html).
-
-## Requirements
-
-To build this sub-project, Java 11+ is required.
-
-## Parser Generation
-
-To build the generated sources, run:
-
-```shell
-./gradlew :partiql-grammar:build
-```
