--- partiql-grammar/build.gradle
@@ -1,33 +1,0 @@
-plugins {
-    id 'antlr'
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    antlr 'org.antlr:antlr4:4.10.1'
-}
-
-java {
-    sourceCompatibility = JavaVersion.VERSION_1_8
-    targetCompatibility = JavaVersion.VERSION_1_8
-}
-
-// ANTLR Configuration
-def ANTLR_INPUT_SOURCE_ROOT = new File(projectDir, "src/main/antlr")
-def ANTLR_OUTPUT_DIR = new File(projectDir, "build/generated-src/antlr/main/org/partiql/grammar/parser/generated")
-
-// ANTLR Task
-generateGrammarSource {
-    maxHeapSize = "64m"
-    arguments += ["-visitor", "-long-messages", "-package", "org.partiql.grammar.parser.generated", "-Xexact-output-dir" ]
-    outputDirectory = ANTLR_OUTPUT_DIR
-}
-
-// Adds G4 Files to Resources
-processResources {
-    from ANTLR_INPUT_SOURCE_ROOT
-    include "**/*.g4"
-}
