+++ settings.gradle
@@ -19,7 +19,6 @@ include('lang',
         'cli', 
         'examples', 
         'extensions',
-        'partiql-grammar',
         'lib:partiql-isl',
         'test:partiql-pts',
         'test:partiql-tests-runner',
